+++ src/Typin/Benchmarks/Typin.Benchmarks.MultiCommand/CliFxCommands/CliFxCommands.cs
@@ -1,6 +1,7 @@
 ﻿namespace Typin.Benchmarks.MultiCommand.CliFxComands
 {
     using CliFx.Attributes;
+    using Typin.Benchmarks.MultiCommand.CliFxCommands;
 
     [Command]
     public class CliFxCommands : CliFxBaseCommand { }