+++ src/Typin/Examples/InteractiveModeExample/Commands/DefaultCommand.cs
@@ -1,8 +1,9 @@
-﻿namespace Typin.InteractiveModeDemo.Commands
+﻿namespace InteractiveModeExample.Commands
 {
     using System;
     using System.Collections.Generic;
     using System.Threading.Tasks;
+    using Typin;
     using Typin.Attributes;
     using Typin.Console;
 
@@ -19,11 +20,10 @@
 
         public ValueTask ExecuteAsync(IConsole console)
         {
-            console.WithForegroundColor(ConsoleColor.DarkGreen, () => console.Output.WriteLine("Hello world from default command"));
+            console.Output.WithForegroundColor(ConsoleColor.DarkGreen, (output) => output.WriteLine("Hello world from default command"));
+
             foreach (var value in Values)
-            {
                 console.Output.WriteLine(value);
-            }
 
             return default;
         }