+++ src/Typin/Typin/Help/DefaultHelpWriter.cs
@@ -1,4 +1,4 @@
-﻿namespace Typin.HelpWriter
+﻿namespace Typin.Help
 {
     using System;
     using System.Collections;
@@ -91,7 +91,7 @@
 
         private void Write(ConsoleColor foregroundColor, string value)
         {
-            _console.WithForegroundColor(foregroundColor, () => Write(value));
+            _console.Output.WithForegroundColor(foregroundColor, (output) => Write(value));
         }
 
         private void WriteLine()