+++ src/Typin/Typin/Utilities/TableUtils.cs
@@ -5,7 +5,7 @@
     using System.Linq;
     using System.Linq.Expressions;
     using Typin.Console;
-    using Typin.Internal.Extensions;
+    using Typin.Extensions;
 
     /// <summary>
     /// Simple table utils for console.
@@ -46,7 +46,7 @@
             console.Output.WriteLine(new string('=', totalWidth));
 
             //Write table header
-            if (headers.Count() > 0)
+            if (headers.Any())
             {
                 for (int i = 0; i < columnsCount; ++i)
                 {
@@ -131,7 +131,7 @@
             console.Output.WriteLine(new string('=', totalWidth));
 
             //Write table header
-            if (headers.Count() > 0)
+            if (headers.Any())
             {
                 for (int i = 0; i < columnsCount; ++i)
                 {